Redesign decryption UI
* Import bootstrap into the iFrame
* Change text display from textarea to div
* Make links clickable (requires escaping input)
* Keep hidden textarea for copying text

Resolves https://github.com/XCF-Babble/babble/issues/48
